Fix request_id attribute in tracing spans
We currently use the http_x_request_id header to set the value for the
request_id span attribute. When the header is not present, the span attribute
will be set to unknown.

The ingress however can generate a random request id in this cases and will
populate the req_id variable with it. It will also properly propagate the value
as a request header to downstream services.

To fix the root span attribute, this commit uses that same variable
instead of accessing the header directly.